+++ plotpy/items/__init__.py
@@ -9,6 +9,7 @@ from .annotation import (
     AnnotatedObliqueRectangle,
     AnnotatedPoint,
     AnnotatedRectangle,
+    AnnotatedPolygon,
     AnnotatedSegment,
     AnnotatedShape,
 )